Split prepare (after fork)
