+++ custom_components/plugwise/manifest.json
@@ -1,13 +1,15 @@
 {
   "domain": "plugwise",
   "name": "Plugwise Beta",
-  "after_dependencies": ["usb", "zeroconf"],
+  "after_dependencies": ["zeroconf"],
   "codeowners": ["@CoMPaTech", "@bouwew", "@brefra"],
   "config_flow": true,
   "documentation": "https://github.com/plugwise/plugwise-beta",
   "integration_type": "hub",
   "iot_class": "local_polling",
   "loggers": ["plugwise"],
-  "requirements": ["plugwise==0.27.10"],
-  "version": "0.34.10"
+  "requirements": [
+    "https://test-files.pythonhosted.org/packages/44/ed/69e2a8587b30162372b843d436a5036c25a3bcdff99c423b4b94666b5de3/plugwise-0.31.0a1.tar.gz#plugwise==0.31.a01"
+  ],
+  "version": "0.40.0a1"
 }
