+++ custom_components/plugwise/models.py
@@ -3,10 +3,7 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 
-from homeassistant.components.binary_sensor import (
-    BinarySensorDeviceClass,
-    BinarySensorEntityDescription,
-)
+from homeassistant.components.binary_sensor import BinarySensorEntityDescription
 from homeassistant.components.sensor import (
     SensorDeviceClass,
     SensorEntityDescription,
@@ -16,20 +13,18 @@ from homeassistant.components.switch import SwitchDeviceClass, SwitchEntityDescr
 from homeassistant.const import (
     LIGHT_LUX,
     PERCENTAGE,
-    SIGNAL_STRENGTH_DECIBELS_MILLIWATT,
     EntityCategory,
     UnitOfElectricPotential,
     UnitOfEnergy,
     UnitOfPower,
     UnitOfPressure,
     UnitOfTemperature,
-    UnitOfTime,
     UnitOfVolume,
     UnitOfVolumeFlowRate,
 )
 from homeassistant.helpers.entity import EntityDescription
 
-from .const import SMILE, STICK, USB_MOTION_ID, USB_RELAY_ID
+from .const import SMILE
 
 
 @dataclass
@@ -77,109 +72,6 @@ class PlugwiseBinarySensorEntityDescription(
 
 
 PW_SENSOR_TYPES: tuple[PlugwiseSensorEntityDescription, ...] = (
-    PlugwiseSensorEntityDescription(
-        plugwise_api=STICK,
-        key="power_1s",
-        name="Power usage",
-        device_class=SensorDeviceClass.POWER,
-        native_unit_of_measurement=UnitOfPower.WATT,
-        state_request_method="current_power_usage",
-    ),
-    PlugwiseSensorEntityDescription(
-        plugwise_api=STICK,
-        key="energy_consumption_today",
-        name="Energy consumption today",
-        device_class=SensorDeviceClass.ENERGY,
-        state_class=SensorStateClass.TOTAL_INCREASING,
-        native_unit_of_measurement=UnitOfEnergy.KILO_WATT_HOUR,
-        state_request_method="energy_consumption_today",
-    ),
-    PlugwiseSensorEntityDescription(
-        plugwise_api=STICK,
-        key="ping",
-        name="Ping roundtrip",
-        icon="mdi:speedometer",
-        native_unit_of_measurement=UnitOfTime.MILLISECONDS,
-        state_request_method="ping",
-        entity_registry_enabled_default=False,
-        entity_category=EntityCategory.DIAGNOSTIC,
-    ),
-    PlugwiseSensorEntityDescription(
-        plugwise_api=STICK,
-        key="power_8s",
-        name="Power usage 8 seconds",
-        device_class=SensorDeviceClass.POWER,
-        native_unit_of_measurement=UnitOfPower.WATT,
-        state_request_method="current_power_usage_8_sec",
-        entity_registry_enabled_default=False,
-    ),
-    PlugwiseSensorEntityDescription(
-        plugwise_api=STICK,
-        key="RSSI_in",
-        name="Inbound RSSI",
-        device_class=SensorDeviceClass.SIGNAL_STRENGTH,
-        native_unit_of_measurement=SIGNAL_STRENGTH_DECIBELS_MILLIWATT,
-        state_request_method="rssi_in",
-        entity_registry_enabled_default=False,
-        entity_category=EntityCategory.DIAGNOSTIC,
-    ),
-    PlugwiseSensorEntityDescription(
-        plugwise_api=STICK,
-        key="RSSI_out",
-        name="Outbound RSSI",
-        device_class=SensorDeviceClass.SIGNAL_STRENGTH,
-        native_unit_of_measurement=SIGNAL_STRENGTH_DECIBELS_MILLIWATT,
-        state_request_method="rssi_out",
-        entity_registry_enabled_default=False,
-        entity_category=EntityCategory.DIAGNOSTIC,
-    ),
-    PlugwiseSensorEntityDescription(
-        plugwise_api=STICK,
-        key="power_con_cur_hour",
-        name="Power consumption current hour",
-        icon="mdi:lightning-bolt",
-        native_unit_of_measurement=UnitOfEnergy.KILO_WATT_HOUR,
-        state_request_method="power_consumption_current_hour",
-        entity_registry_enabled_default=False,
-    ),
-    PlugwiseSensorEntityDescription(
-        plugwise_api=STICK,
-        key="power_prod_cur_hour",
-        name="Power production current hour",
-        icon="mdi:lightning-bolt",
-        native_unit_of_measurement=UnitOfEnergy.KILO_WATT_HOUR,
-        state_request_method="power_production_current_hour",
-        entity_registry_enabled_default=False,
-    ),
-    PlugwiseSensorEntityDescription(
-        plugwise_api=STICK,
-        key="power_con_today",
-        name="Power consumption today",
-        icon="mdi:lightning-bolt",
-        native_unit_of_measurement=UnitOfEnergy.KILO_WATT_HOUR,
-        state_request_method="power_consumption_today",
-        entity_registry_enabled_default=False,
-    ),
-    PlugwiseSensorEntityDescription(
-        plugwise_api=STICK,
-        key="power_con_prev_hour",
-        name="Power consumption previous hour",
-        icon="mdi:lightning-bolt",
-        native_unit_of_measurement=UnitOfEnergy.KILO_WATT_HOUR,
-        state_class=SensorStateClass.TOTAL,
-        state_request_method="power_consumption_previous_hour",
-        entity_registry_enabled_default=False,
-    ),
-    PlugwiseSensorEntityDescription(
-        plugwise_api=STICK,
-        key="power_con_yesterday",
-        name="Power consumption yesterday",
-        icon="mdi:lightning-bolt",
-        native_unit_of_measurement=UnitOfEnergy.KILO_WATT_HOUR,
-        state_class=SensorStateClass.TOTAL,
-        state_request_method="power_consumption_yesterday",
-        entity_registry_enabled_default=False,
-    ),
     PlugwiseSensorEntityDescription(
         plugwise_api=SMILE,
         key="setpoint",
@@ -566,13 +458,6 @@ PW_SENSOR_TYPES: tuple[PlugwiseSensorEntityDescription, ...] = (
 
 PW_SWITCH_TYPES: tuple[PlugwiseSwitchEntityDescription, ...] = (
     PlugwiseSwitchEntityDescription(
-        plugwise_api=STICK,
-        key=USB_RELAY_ID,
-        device_class=SwitchDeviceClass.OUTLET,
-        name="Relay state",
-        state_request_method="relay_state",
-    ),
-    PlugwiseSwitchEntityDescription(
         plugwise_api=SMILE,
         key="dhw_cm_switch",
         translation_key="dhw_cm_switch",
@@ -605,13 +490,6 @@ PW_SWITCH_TYPES: tuple[PlugwiseSwitchEntityDescription, ...] = (
 )
 
 PW_BINARY_SENSOR_TYPES: tuple[PlugwiseBinarySensorEntityDescription, ...] = (
-    PlugwiseBinarySensorEntityDescription(
-        plugwise_api=STICK,
-        key=USB_MOTION_ID,
-        name="Motion",
-        device_class=BinarySensorDeviceClass.MOTION,
-        state_request_method="motion",
-    ),
     PlugwiseBinarySensorEntityDescription(
         plugwise_api=SMILE,
         key="compressor_state",
