+++ tests/components/plugwise/test_config_flow.py
@@ -2,23 +2,13 @@
 from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
-import serial.tools.list_ports
-from voluptuous.error import MultipleInvalid
 
 from homeassistant.components import zeroconf
-from homeassistant.components.plugwise.config_flow import CONF_MANUAL_PATH
 from homeassistant.components.plugwise.const import (
-    API,
     CONF_HOMEKIT_EMULATION,
     CONF_REFRESH_INTERVAL,
-    CONF_USB_PATH,
     DEFAULT_PORT,
     DOMAIN,
-    FLOW_NET,
-    FLOW_TYPE,
-    FLOW_USB,
-    PW_TYPE,
-    STICK,
 )
 from homeassistant.components.zeroconf import ZeroconfServiceInfo
 from homeassistant.config_entries import SOURCE_USER, SOURCE_ZEROCONF
@@ -38,10 +28,6 @@ from plugwise.exceptions import (
     InvalidAuthentication,
     InvalidSetupError,
     InvalidXMLError,
-    NetworkDown,
-    ResponseError,
-    StickInitError,
-    TimeoutException,
     UnsupportedDeviceError,
 )
 
@@ -52,8 +38,6 @@ TEST_HOSTNAME = "smileabcdef"
 TEST_HOSTNAME2 = "stretchabc"
 TEST_PASSWORD = "test_password"
 TEST_PORT = 81
-TEST_USBPORT = "/dev/ttyUSB1"
-TEST_USBPORT2 = "/dev/ttyUSB2"
 TEST_USERNAME = "smile"
 TEST_USERNAME2 = "stretch"
 
@@ -123,64 +107,9 @@ def mock_smile():
         smile_mock.ConnectionFailedError = ConnectionFailedError
         smile_mock.InvalidAuthentication = InvalidAuthentication
         smile_mock.InvalidXMLError = InvalidXMLError
-        smile_mock.ResponseError = ResponseError
         smile_mock.UnsupportedDeviceError = UnsupportedDeviceError
         smile_mock.return_value.connect.return_value = True
         yield smile_mock.return_value
-
-
-def com_port():
-    """Mock of a serial port."""
-    port = serial.tools.list_ports_common.ListPortInfo(TEST_USBPORT)
-    port.serial_number = "1234"
-    port.manufacturer = "Virtual serial port"
-    port.device = TEST_USBPORT
-    port.description = "Some serial port"
-    return port
-
-
-async def test_form_flow_gateway(
-    hass: HomeAssistant,
-    mock_setup_entry: AsyncMock,
-    mock_smile_config_flow: MagicMock,
-) -> None:
-    """Test we get the form for Plugwise Gateway product type."""
-    result = await hass.config_entries.flow.async_init(
-        DOMAIN, context={CONF_SOURCE: SOURCE_USER}
-    )
-    assert result.get("type") == FlowResultType.FORM
-    assert result.get("errors") == {}
-    assert result.get("step_id") == "user"
-    assert "flow_id" in result
-
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"], user_input={FLOW_TYPE: FLOW_NET}
-    )
-    assert result.get("type") == FlowResultType.FORM
-    assert result.get("errors") == {}
-    assert result.get("step_id") == "user_gateway"
-
-
-async def test_form_flow_usb(
-    hass: HomeAssistant,
-    mock_setup_entry: AsyncMock,
-    mock_smile_config_flow: MagicMock,
-) -> None:
-    """Test we get the form for Plugwise USB product type."""
-    result = await hass.config_entries.flow.async_init(
-        DOMAIN, context={CONF_SOURCE: SOURCE_USER}
-    )
-    assert result.get("type") == FlowResultType.FORM
-    assert result.get("errors") == {}
-    assert result.get("step_id") == "user"
-    assert "flow_id" in result
-
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"], user_input={FLOW_TYPE: FLOW_USB}
-    )
-    assert result.get("type") == FlowResultType.FORM
-    assert result.get("errors") == {}
-    assert result.get("step_id") == "user_usb"
 
 
 async def test_form(
@@ -190,12 +119,11 @@ async def test_form(
 ) -> None:
     """Test the full user configuration flow."""
     result = await hass.config_entries.flow.async_init(
-        DOMAIN, context={CONF_SOURCE: SOURCE_USER}, data={FLOW_TYPE: FLOW_NET}
+        DOMAIN, context={CONF_SOURCE: SOURCE_USER}
     )
     assert result.get("type") == FlowResultType.FORM
     assert result.get("errors") == {}
-    assert result.get("step_id") == "user_gateway"
-    assert "flow_id" in result
+    assert result.get("step_id") == "user"
 
     result2 = await hass.config_entries.flow.async_configure(
         result["flow_id"],
@@ -213,7 +141,6 @@ async def test_form(
         CONF_PASSWORD: TEST_PASSWORD,
         CONF_PORT: DEFAULT_PORT,
         CONF_USERNAME: TEST_USERNAME,
-        PW_TYPE: API,
     }
 
     assert len(mock_setup_entry.mock_calls) == 1
@@ -242,7 +169,7 @@ async def test_zeroconf_form(
     )
     assert result.get("type") == FlowResultType.FORM
     assert result.get("errors") == {}
-    assert result.get("step_id") == "user_gateway"
+    assert result.get("step_id") == "user"
     assert "flow_id" in result
 
     result2 = await hass.config_entries.flow.async_configure(
@@ -258,7 +185,6 @@ async def test_zeroconf_form(
         CONF_PASSWORD: TEST_PASSWORD,
         CONF_PORT: DEFAULT_PORT,
         CONF_USERNAME: TEST_USERNAME,
-        PW_TYPE: API,
     }
 
     assert len(mock_setup_entry.mock_calls) == 1
@@ -278,7 +204,7 @@ async def test_zeroconf_stretch_form(
     )
     assert result.get("type") == FlowResultType.FORM
     assert result.get("errors") == {}
-    assert result.get("step_id") == "user_gateway"
+    assert result.get("step_id") == "user"
     assert "flow_id" in result
 
     result2 = await hass.config_entries.flow.async_configure(
@@ -294,7 +220,6 @@ async def test_zeroconf_stretch_form(
         CONF_PASSWORD: TEST_PASSWORD,
         CONF_PORT: DEFAULT_PORT,
         CONF_USERNAME: TEST_USERNAME2,
-        PW_TYPE: API,
     }
 
     assert len(mock_setup_entry.mock_calls) == 1
@@ -324,7 +249,7 @@ async def test_zeroconf_abort_anna_with_adam(hass: HomeAssistant) -> None:
         data=TEST_DISCOVERY_ANNA,
     )
     assert result.get("type") == FlowResultType.FORM
-    assert result.get("step_id") == "user_gateway"
+    assert result.get("step_id") == "user"
 
     flows_in_progress = hass.config_entries.flow.async_progress()
     assert len(flows_in_progress) == 1
@@ -338,7 +263,7 @@ async def test_zeroconf_abort_anna_with_adam(hass: HomeAssistant) -> None:
     )
 
     assert result2.get("type") == FlowResultType.FORM
-    assert result2.get("step_id") == "user_gateway"
+    assert result2.get("step_id") == "user"
 
     flows_in_progress = hass.config_entries.flow.async_progress()
     assert len(flows_in_progress) == 1
@@ -409,7 +334,6 @@ async def test_zercoconf_discovery_update_configuration(
         (InvalidAuthentication, "invalid_auth"),
         (InvalidSetupError, "invalid_setup"),
         (InvalidXMLError, "response_error"),
-        (ResponseError, "response_error"),
         (RuntimeError, "unknown"),
         (UnsupportedDeviceError, "unsupported"),
     ],
@@ -433,16 +357,13 @@ async def test_flow_errors(
 
     mock_smile_config_flow.connect.side_effect = side_effect
     result2 = await hass.config_entries.flow.async_configure(
-        result["flow_id"], user_input={FLOW_TYPE: FLOW_NET}
-    )
-    result3 = await hass.config_entries.flow.async_configure(
-        result2["flow_id"],
+        result["flow_id"],
         user_input={CONF_HOST: TEST_HOST, CONF_PASSWORD: TEST_PASSWORD},
     )
 
-    assert result3.get("type") == FlowResultType.FORM
-    assert result3.get("errors") == {"base": reason}
-    assert result3.get("step_id") == "user_gateway"
+    assert result2.get("type") == FlowResultType.FORM
+    assert result2.get("errors") == {"base": reason}
+    assert result2.get("step_id") == "user"
 
     assert len(mock_setup_entry.mock_calls) == 0
     assert len(mock_smile_config_flow.connect.mock_calls) == 1
@@ -460,7 +381,6 @@ async def test_flow_errors(
         CONF_PASSWORD: TEST_PASSWORD,
         CONF_PORT: DEFAULT_PORT,
         CONF_USERNAME: TEST_USERNAME,
-        PW_TYPE: API,
     }
 
     assert len(mock_setup_entry.mock_calls) == 1
@@ -470,7 +390,8 @@ async def test_flow_errors(
 async def test_form_invalid_setup(hass, mock_smile):
     """Test we handle invalid setup."""
     result = await hass.config_entries.flow.async_init(
-        DOMAIN, context={CONF_SOURCE: SOURCE_USER}, data={FLOW_TYPE: FLOW_NET}
+        DOMAIN,
+        context={CONF_SOURCE: SOURCE_USER},
     )
 
     mock_smile.connect.side_effect = InvalidSetupError
@@ -488,7 +409,8 @@ async def test_form_invalid_setup(hass, mock_smile):
 async def test_form_invalid_auth(hass, mock_smile):
     """Test we handle invalid auth."""
     result = await hass.config_entries.flow.async_init(
-        DOMAIN, context={CONF_SOURCE: SOURCE_USER}, data={FLOW_TYPE: FLOW_NET}
+        DOMAIN,
+        context={CONF_SOURCE: SOURCE_USER},
     )
 
     mock_smile.connect.side_effect = InvalidAuthentication
@@ -506,7 +428,8 @@ async def test_form_invalid_auth(hass, mock_smile):
 async def test_form_cannot_connect(hass, mock_smile):
     """Test we handle cannot connect error."""
     result = await hass.config_entries.flow.async_init(
-        DOMAIN, context={CONF_SOURCE: SOURCE_USER}, data={FLOW_TYPE: FLOW_NET}
+        DOMAIN,
+        context={CONF_SOURCE: SOURCE_USER},
     )
 
     mock_smile.connect.side_effect = ConnectionFailedError
@@ -524,7 +447,8 @@ async def test_form_cannot_connect(hass, mock_smile):
 async def test_form_cannot_connect_port(hass, mock_smile):
     """Test we handle cannot connect to port error."""
     result = await hass.config_entries.flow.async_init(
-        DOMAIN, context={CONF_SOURCE: SOURCE_USER}, data={FLOW_TYPE: FLOW_NET}
+        DOMAIN,
+        context={CONF_SOURCE: SOURCE_USER},
     )
 
     mock_smile.connect.side_effect = ConnectionFailedError
@@ -546,7 +470,8 @@ async def test_form_cannot_connect_port(hass, mock_smile):
 async def test_form_other_problem(hass, mock_smile):
     """Test we handle cannot connect error."""
     result = await hass.config_entries.flow.async_init(
-        DOMAIN, context={CONF_SOURCE: SOURCE_USER}, data={FLOW_TYPE: FLOW_NET}
+        DOMAIN,
+        context={CONF_SOURCE: SOURCE_USER},
     )
 
     mock_smile.connect.side_effect = TimeoutError
@@ -598,224 +523,3 @@ async def test_options_flow_thermo(hass, mock_smile_anna_2) -> None:
             CONF_REFRESH_INTERVAL: 3.0,
             CONF_SCAN_INTERVAL: 60,
         }
-
-
-@patch("serial.tools.list_ports.comports", MagicMock(return_value=[com_port()]))
-@patch("plugwise.Stick.connect", MagicMock(return_value=None))
-@patch("plugwise.Stick.initialize_stick", MagicMock(return_value=None))
-@patch("plugwise.Stick.disconnect", MagicMock(return_value=None))
-async def test_user_flow_select(hass):
-    """Test user flow when USB-stick is selected from list."""
-    port = com_port()
-    port_select = f"{port}, s/n: {port.serial_number} - {port.manufacturer}"
-
-    result = await hass.config_entries.flow.async_init(
-        DOMAIN, context={CONF_SOURCE: SOURCE_USER}, data={FLOW_TYPE: FLOW_USB}
-    )
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"], user_input={CONF_USB_PATH: port_select}
-    )
-    assert result["type"] == FlowResultType.CREATE_ENTRY
-    assert result["data"] == {PW_TYPE: STICK, CONF_USB_PATH: TEST_USBPORT}
-
-    # Retry to ensure configuring the same port is not allowed
-    result = await hass.config_entries.flow.async_init(
-        DOMAIN, context={CONF_SOURCE: SOURCE_USER}, data={FLOW_TYPE: FLOW_USB}
-    )
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"], user_input={CONF_USB_PATH: port_select}
-    )
-    assert result["type"] == FlowResultType.FORM
-    assert result["errors"] == {"base": "already_configured"}
-
-
-async def test_user_flow_manual_selected_show_form(hass):
-    """Test user step form when manual path is selected."""
-    result = await hass.config_entries.flow.async_init(
-        DOMAIN, context={CONF_SOURCE: SOURCE_USER}, data={FLOW_TYPE: FLOW_USB}
-    )
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        user_input={CONF_USB_PATH: CONF_MANUAL_PATH},
-    )
-    assert result["type"] == FlowResultType.FORM
-    assert result["step_id"] == "manual_path"
-
-
-async def test_user_flow_manual(hass):
-    """Test user flow when USB-stick is manually entered."""
-    result = await hass.config_entries.flow.async_init(
-        DOMAIN, context={CONF_SOURCE: SOURCE_USER}, data={FLOW_TYPE: FLOW_USB}
-    )
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        user_input={CONF_USB_PATH: CONF_MANUAL_PATH},
-    )
-
-    with patch(
-        "homeassistant.components.plugwise.config_flow.Stick",
-    ) as usb_mock:
-        usb_mock.return_value.connect = MagicMock(return_value=True)
-        usb_mock.return_value.initialize_stick = MagicMock(return_value=True)
-        usb_mock.return_value.disconnect = MagicMock(return_value=True)
-        usb_mock.return_value.mac = "01:23:45:67:AB"
-
-        result = await hass.config_entries.flow.async_configure(
-            result["flow_id"],
-            user_input={CONF_USB_PATH: TEST_USBPORT2},
-        )
-    assert result["type"] == FlowResultType.CREATE_ENTRY
-    assert result["data"] == {CONF_USB_PATH: TEST_USBPORT2}
-
-
-async def test_invalid_connection(hass):
-    """Test invalid connection."""
-    result = await hass.config_entries.flow.async_init(
-        DOMAIN, context={CONF_SOURCE: SOURCE_USER}, data={FLOW_TYPE: FLOW_USB}
-    )
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        user_input={CONF_USB_PATH: CONF_MANUAL_PATH},
-    )
-
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        {CONF_USB_PATH: "/dev/null"},
-    )
-
-    assert result["type"] == FlowResultType.FORM
-    assert result["errors"] == {"base": "cannot_connect"}
-
-
-async def test_empty_connection(hass):
-    """Test empty connection."""
-    result = await hass.config_entries.flow.async_init(
-        DOMAIN, context={CONF_SOURCE: SOURCE_USER}, data={FLOW_TYPE: FLOW_USB}
-    )
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        user_input={CONF_USB_PATH: CONF_MANUAL_PATH},
-    )
-
-    try:
-        result = await hass.config_entries.flow.async_configure(
-            result["flow_id"],
-            {CONF_USB_PATH: None},
-        )
-        pytest.fail("Empty connection was accepted")
-    except MultipleInvalid:
-        assert True
-
-    assert result["type"] == FlowResultType.FORM
-    assert result["errors"] == {}
-
-
-@patch("plugwise.Stick.connect", MagicMock(return_value=None))
-@patch("plugwise.Stick.initialize_stick", MagicMock(side_effect=(StickInitError)))
-async def test_failed_initialization(hass):
-    """Test we handle failed initialization of Plugwise USB-stick."""
-    result = await hass.config_entries.flow.async_init(
-        DOMAIN,
-        context={CONF_SOURCE: SOURCE_USER},
-        data={FLOW_TYPE: FLOW_USB},
-    )
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        user_input={CONF_USB_PATH: CONF_MANUAL_PATH},
-    )
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        user_input={CONF_USB_PATH: "/dev/null"},
-    )
-    assert result["type"] == "form"
-    assert result["errors"] == {"base": "stick_init"}
-
-
-@patch("plugwise.Stick.connect", MagicMock(return_value=None))
-@patch("plugwise.Stick.initialize_stick", MagicMock(side_effect=(NetworkDown)))
-async def test_network_down_exception(hass):
-    """Test we handle network_down exception."""
-    result = await hass.config_entries.flow.async_init(
-        DOMAIN,
-        context={CONF_SOURCE: SOURCE_USER},
-        data={FLOW_TYPE: FLOW_USB},
-    )
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        user_input={CONF_USB_PATH: CONF_MANUAL_PATH},
-    )
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        user_input={CONF_USB_PATH: "/dev/null"},
-    )
-    assert result["type"] == "form"
-    assert result["errors"] == {"base": "network_down"}
-
-
-@patch("plugwise.Stick.connect", MagicMock(return_value=None))
-@patch("plugwise.Stick.initialize_stick", MagicMock(side_effect=(TimeoutException)))
-async def test_timeout_exception(hass):
-    """Test we handle time exception."""
-    result = await hass.config_entries.flow.async_init(
-        DOMAIN,
-        context={CONF_SOURCE: SOURCE_USER},
-        data={FLOW_TYPE: FLOW_USB},
-    )
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        user_input={CONF_USB_PATH: CONF_MANUAL_PATH},
-    )
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"],
-        user_input={CONF_USB_PATH: "/dev/null"},
-    )
-    assert result["type"] == "form"
-    assert result["errors"] == {"base": "network_timeout"}
-
-
-async def test_options_flow_stick(hass) -> None:
-    """Test config flow options lack for stick environments."""
-    entry = MockConfigEntry(
-        domain=DOMAIN,
-        title=CONF_NAME,
-        data={FLOW_TYPE: FLOW_USB},
-    )
-    hass.data[DOMAIN] = {entry.entry_id: {"api_stick": MagicMock()}}
-    entry.add_to_hass(hass)
-
-    with patch(
-        "homeassistant.components.plugwise.async_setup_entry", return_value=True
-    ):
-        assert await hass.config_entries.async_setup(entry.entry_id)
-        await hass.async_block_till_done()
-
-        result = await hass.config_entries.options.async_init(entry.entry_id)
-
-        assert result["type"] == FlowResultType.FORM
-        assert result["step_id"] == "none"
-
-
-async def test_options_flow_stick_with_input(hass) -> None:
-    """Test config flow options lack for stick environments."""
-    entry = MockConfigEntry(
-        domain=DOMAIN,
-        title=CONF_NAME,
-        data={FLOW_TYPE: FLOW_USB},
-    )
-    hass.data[DOMAIN] = {entry.entry_id: {"api_stick": MagicMock()}}
-    entry.add_to_hass(hass)
-
-    with patch(
-        "homeassistant.components.plugwise.async_setup_entry", return_value=True
-    ):
-        assert await hass.config_entries.async_setup(entry.entry_id)
-        await hass.async_block_till_done()
-
-        result = await hass.config_entries.options.async_init(entry.entry_id)
-        result = await hass.config_entries.options.async_configure(
-            result["flow_id"],
-            user_input={CONF_USB_PATH: TEST_USBPORT2},
-        )
-
-        assert result["type"] == FlowResultType.CREATE_ENTRY
-        assert result["title"] == ""
